+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+- Fix handling of `completionItem/resolve` when not all fields are present on the `CompletionItem`.
+- Fix handling of eager resolution of completions.
+
 ## 0.25.4
 
 ### Fixed